test: fix tests discovery
